requirements.txt
# Telegram
python-telegram-bot==21.7

# LLM (Sambanova — API compatible con OpenAI)
openai>=1.0.0

# Embeddings locales
sentence-transformers>=3.0.0

# RAG
chromadb>=0.5.0

# Scheduler
APScheduler>=3.10
pytz

# PDF support (opcional, para guías en PDF)
PyMuPDF>=1.24.0

# Env
python-dotenv>=1.0
